requirements.txt
discord.py>=2.7.1,<3.0.0
aiohttp>=3.13.3,<4.0.0
matplotlib>=3.10.0,<4.0.0
python-dotenv>=1.2.2,<2.0.0
